move cat.js => unit.mjs
