+++ .gitattributes
@@ -5,6 +5,7 @@
 *.txt text eol=lf
 *.json text eol=lf
 *.js text eol=lf
+*.mjs text eol=lf
 *.tsv text eol=lf
 *.csv text eol=lf
 *.html text eol=lf
